backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
python-dotenv
pandas
openai
